@@ -133,7 +133,8 @@ namespace RabbitMQ.Client.Framing.Impl
                 byte[]? challenge = null;
                 do
                 {
-                    byte[] response = mechanism.handleChallenge(challenge, _config);
+                    byte[] response = await mechanism.HandleChallengeAsync(challenge, _config)
+                        .ConfigureAwait(false);
                     ConnectionSecureOrTune res;
                     if (challenge is null)
                     {
@@ -183,29 +184,10 @@ namespace RabbitMQ.Client.Framing.Impl
                 .ConfigureAwait(false);
 
             cancellationToken.ThrowIfCancellationRequested();
-            MaybeStartCredentialRefresher();
 
             // now we can start heartbeat timers
             cancellationToken.ThrowIfCancellationRequested();
             MaybeStartHeartbeatTimers();
-        }
-
-        private void MaybeStartCredentialRefresher()
-        {
-            if (_config.CredentialsProvider.ValidUntil != null)
-            {
-                _config.CredentialsRefresher.Register(_config.CredentialsProvider, NotifyCredentialRefreshedAsync);
-            }
-        }
-
-        private async Task NotifyCredentialRefreshedAsync(bool succesfully)
-        {
-            if (succesfully)
-            {
-                using var cts = new CancellationTokenSource(InternalConstants.DefaultConnectionCloseTimeout);
-                await UpdateSecretAsync(_config.CredentialsProvider.Password, "Token refresh", cts.Token)
-                    .ConfigureAwait(false);
-            }
         }
 
         private IAuthMechanismFactory GetAuthMechanismFactory(string supportedMechanismNames)
